+++ druid-shell/src/platform/mac/mod.rs
@@ -25,3 +25,4 @@ mod keyboard;
 pub mod menu;
 pub mod util;
 pub mod window;
+pub mod screen;