+++ druid-shell/src/platform/windows/application.rs
@@ -22,14 +22,14 @@ use std::rc::Rc;
 
 use winapi::shared::minwindef::{FALSE, HINSTANCE};
 use winapi::shared::ntdef::LPCWSTR;
-use winapi::shared::windef::{HCURSOR, HWND};
+use winapi::shared::windef::{HCURSOR, HWND, DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2};
 use winapi::shared::winerror::HRESULT_FROM_WIN32;
 use winapi::um::errhandlingapi::GetLastError;
-use winapi::um::shellscalingapi::PROCESS_SYSTEM_DPI_AWARE;
+use winapi::um::shellscalingapi::PROCESS_PER_MONITOR_DPI_AWARE;
 use winapi::um::winuser::{
     DispatchMessageW, GetAncestor, GetMessageW, LoadIconW, PostMessageW, PostQuitMessage,
     RegisterClassW, TranslateAcceleratorW, TranslateMessage, GA_ROOT, IDI_APPLICATION, MSG,
-    WNDCLASSW,
+    WNDCLASSW, SendMessageW
 };
 
 use crate::application::AppHandler;
@@ -38,7 +38,7 @@ use super::accels;
 use super::clipboard::Clipboard;
 use super::error::Error;
 use super::util::{self, ToWide, CLASS_NAME, OPTIONAL_FUNCTIONS};
-use super::window::{self, DS_REQUEST_DESTROY};
+use super::window::{self, DS_REQUEST_DESTROY, DS_HANDLE_DEFERRED};
 
 #[derive(Clone)]
 pub(crate) struct Application {
@@ -49,6 +49,8 @@ struct State {
     quitting: bool,
     windows: HashSet<HWND>,
 }
+
+
 
 impl Application {
     pub fn new() -> Result<Application, Error> {
@@ -64,10 +66,15 @@ impl Application {
     fn init() -> Result<(), Error> {
         // TODO: Report back an error instead of panicking
         util::attach_console();
-        if let Some(func) = OPTIONAL_FUNCTIONS.SetProcessDpiAwareness {
+        if let Some(func) = OPTIONAL_FUNCTIONS.SetProcessDpiAwarenessContext {
             // This function is only supported on windows 10
             unsafe {
-                func(PROCESS_SYSTEM_DPI_AWARE); // TODO: per monitor (much harder)
+                func(DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2);
+            }
+        }
+        else if let Some(func) = OPTIONAL_FUNCTIONS.SetProcessDpiAwareness {
+            unsafe {
+                func(PROCESS_PER_MONITOR_DPI_AWARE);
             }
         }
         unsafe {
@@ -105,6 +112,11 @@ impl Application {
         unsafe {
             // Handle windows messages
             loop {
+                if let Ok(state) = self.state.try_borrow() {
+                    for hwnd in &state.windows {
+                        SendMessageW(*hwnd, DS_HANDLE_DEFERRED, 0,0);
+                    }
+                }
                 let mut msg = mem::MaybeUninit::uninit();
                 let res = GetMessageW(msg.as_mut_ptr(), ptr::null_mut(), 0, 0);
                 if res <= 0 {