+++ druid/src/lib.rs
@@ -179,7 +179,8 @@ pub use shell::keyboard_types;
 pub use shell::{
     Application, Clipboard, ClipboardFormat, Code, Cursor, Error as PlatformError,
     FileDialogOptions, FileInfo, FileSpec, FormatId, HotKey, KbKey, KeyEvent, Location, Modifiers,
-    MouseButton, MouseButtons, RawMods, Region, Scalable, Scale, SysMods, TimerToken, WindowHandle,
+    Monitor, MouseButton, MouseButtons, RawMods, Region, Scalable, Scale, Screen, SysMods,
+    TimerToken, WindowHandle, WindowState,
 };
 
 pub use crate::core::WidgetPod;